+++ src/app/(root)/join/page.tsx
