+++ src/components/ui/navber-mobile.tsx
@@ -68,13 +68,11 @@ const NavbarMobile = () => {
 
                     <div className="flex flex-col items-center">
                         <Link href="/" className="text-white text-lg mb-6">Home</Link>
-                        {/* <Link href="/about" className="text-white text-lg mb-6">Contact Us</Link> */}
-                        <Link href="/about" className="text-white text-lg mb-6">About Us</Link>
-                        <Link href="/sessions" className="text-white text-lg mb-6">Sessions</Link>
+                        <Link href="/contact" className="text-white text-lg mb-6">Contact</Link>                        
                         <Link href="/fests" className="text-white text-lg mb-6">Fests</Link>
                         <Link href="/members" className="text-white text-lg mb-6">Members</Link>
                         <Link href="/events" className="text-white text-lg mb-6">Events</Link>
-                        <Link href="/form" className="text-white text-lg mb-6">Form</Link>
+                        <Link href="/form" className="text-white text-lg mb-6">Join</Link>
                     </div>
                 </div>
             )}
